Turtle.Internal: bug fix: closeCurrent now produces only one new path
The problem was the call to 'setPenPos', which not only reset the pen
position but also created a new output path from the current path;
closeCurrent then proceeded to add a closed version of that path as
well. I.e. the problem is that 'setPenPos' is a 'smart' function made
for consumption by users of the API; all we want to do in closeCurrent
is *literally* set the pen position, and nothing else.  So, now that's
exactly what we do, which makes the code simpler anyway.

Fixes #13.