avoid `hiding` in imports
495c146f552e6832ec9351b3e21639ae05cc468d hides the import of Empty from
`lens` to avoid name collision with `lens-4.13` and `GHC-7.10`.
Unfortunately, this produces a compiler warning with earlier lens or
GHC, which do not export Empty.  The warning becomes an error in our CI
builds.

This commit explicitly imports the needed modules, and the handful of
names from `lens`, which should be less fragile.